requirements.txt
paramiko>=2.7.0
PyYAML>=5.1
tqdm>=4.45.0
PyQt6>=6.4.0
pyinstaller>=5.6.0
pillow>=8.0.0 